fix: change type from double to long fix: remove Math.Round() from MemoryMetricsService fix: change MBytes to Bytes

+++ InstanceManager/Entities/MemoryMetrics.cs
@@ -2,7 +2,7 @@
 
 public class MemoryMetrics
 {
-    public double Total;
-    public double Used;
-    public double Free;
+    public long Total;
+    public long Used;
+    public long Free;
 }
